requirements-test.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0
pytest-playwright>=0.4.0
pytest-mock>=3.10.0
fastapi>=0.104.0
apscheduler>=3.10.0
